refactor: change dir name from lib to init
