+++ init/swagger/core/ApiRequestOptions.ts
