+++ init/swagger/core/BaseHttpRequest.ts
