+++ init/swagger/models/CategoryEnum.ts
