+++ init/swagger/models/PatchedUser.ts
