+++ init/swagger/models/ProviderAuth.ts
