+++ package.json
@@ -24,7 +24,7 @@
     "fix:style": "stylelint --fix '**/*.{css,scss,sass}'",
     "test": "jest --",
     "test:watch": "jest --watch",
-    "swagger": "openapi --input http://localhost:8888/swagger/ --output ./lib/swagger/ --client axios --name SwaggerClient",
+    "swagger": "openapi --input http://localhost:8888/swagger/ --output ./init/swagger/ --client axios --name SwaggerClient",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook",
     "serve-storybook": "serve storybook-static"
