+++ pages/_app.tsx
@@ -7,9 +7,9 @@ import { ThemeProvider } from '@mui/material/styles';
 import CssBaseline from '@mui/material/CssBaseline';
 import { CacheProvider, EmotionCache } from '@emotion/react';
 import theme from '../styles/initAppTheme';
-import { createEmotionCache } from '@/lib/emotion';
+import { createEmotionCache } from 'init/emotion';
 import { SWRConfig } from 'swr';
-import { fetcher } from '../lib/axios';
+import { fetcher } from '../init/axios';
 import { AuthProvider } from '@/providers/AuthProvider';
 import { RouteGuardProvider } from '@/providers/RouteGuardProvider';
 import Backdrop from '@mui/material/Backdrop';
