+++ pages/showcase/forms/ValidatedFormExample.tsx
@@ -13,7 +13,7 @@ import {
 import DeleteIcon from '@mui/icons-material/Delete';
 import { useForm } from 'react-hook-form';
 import { RHFSelect } from '@/components/select/RHFSelect';
-import yup from '@/lib/yup';
+import yup from 'init/yup';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { useSnackbar } from 'hooks/useSnackbar';
 import BrothSnackbar from '@/components/snackbar/BrothSnackbar';
